+++ example/models.g.dart
@@ -18,15 +18,17 @@ const $$Person_fields_name = const DeclarationMirror(
     type: String, annotations: const [const MyAnnotation('hello', val2: null)]);
 const $$Person_fields_myDynamic = const DeclarationMirror(type: dynamic);
 const $$Person_fields_cars = const DeclarationMirror(type: const [List, Car]);
-const $$Person_fields_myGetter = const DeclarationMirror(type: String);
+const $$Person_fields_myGetter =
+    const DeclarationMirror(type: String, isFinal: true);
 const $$Person_fields_mySetter = const DeclarationMirror(type: String);
 
-const PersonClassMirror = const ClassMirror(constructors: const {
+const PersonClassMirror =
+    const ClassMirror(name: 'Person', constructors: const {
   '': const FunctionMirror(parameters: const {
-    'id': const DeclarationMirror(type: int),
-    'name': const DeclarationMirror(type: String),
-    'myDynamic': const DeclarationMirror(type: dynamic),
-    'cars': const DeclarationMirror(type: const [List, Car])
+    'id': const DeclarationMirror(type: int, isOptional: true),
+    'name': const DeclarationMirror(type: String, isOptional: true),
+    'myDynamic': const DeclarationMirror(type: dynamic, isOptional: true),
+    'cars': const DeclarationMirror(type: const [List, Car], isOptional: true)
   }, call: _Person__Constructor)
 }, fields: const {
   'id': $$Person_fields_id,
@@ -59,10 +61,10 @@ _Car__Constructor(params) => new Car(params['id'], params['engine']);
 const $$Car_fields_id = const DeclarationMirror(type: int);
 const $$Car_fields_engine = const DeclarationMirror(type: String);
 
-const CarClassMirror = const ClassMirror(constructors: const {
+const CarClassMirror = const ClassMirror(name: 'Car', constructors: const {
   '': const FunctionMirror(parameters: const {
-    'id': const DeclarationMirror(type: int),
-    'engine': const DeclarationMirror(type: String)
+    'id': const DeclarationMirror(type: int, isOptional: true),
+    'engine': const DeclarationMirror(type: String, isOptional: true)
   }, call: _Car__Constructor)
 }, annotations: const [
   myOtherAnnotation
@@ -84,7 +86,8 @@ const CarClassMirror = const ClassMirror(constructors: const {
 
 _EmptyClass__Constructor(params) => new EmptyClass();
 
-const EmptyClassClassMirror = const ClassMirror(constructors: const {
+const EmptyClassClassMirror =
+    const ClassMirror(name: 'EmptyClass', constructors: const {
   '': const FunctionMirror(parameters: const {}, call: _EmptyClass__Constructor)
 });
 
@@ -100,7 +103,8 @@ const $$ExtendedPerson_fields_extendedName =
 const $$ExtendedPerson_fields_otherExtended =
     const DeclarationMirror(type: dynamic);
 
-const ExtendedPersonClassMirror = const ClassMirror(constructors: const {
+const ExtendedPersonClassMirror =
+    const ClassMirror(name: 'ExtendedPerson', constructors: const {
   '': const FunctionMirror(
       parameters: const {}, call: _ExtendedPerson__Constructor)
 }, fields: const {