+++ lib/built_mirrors.dart
@@ -27,6 +27,14 @@ abstract class Annotation {
 
 /// reflects a Dart language class.
 class ClassMirror {
+  /// name of the class
+  final String name;
+
+  /// If the class is enum
+  final bool isEnum;
+
+  /// If the class mirror is enum  then this will return the values of the enum
+  final List values;
   final List<Annotation> annotations;
   final Map<String, FunctionMirror> constructors;
   final Map<String, DeclarationMirror> fields;
@@ -35,6 +43,9 @@ class ClassMirror {
   final Map<String, FunctionMirror> methods;
 
   const ClassMirror({
+    this.name,
+    this.isEnum = false,
+    this.values,
     this.annotations,
     this.constructors,
     this.fields,
@@ -42,11 +53,15 @@ class ClassMirror {
     this.setters,
     this.methods
   });
+
+  toString() => 'ClassMirror on $name';
 }
+
+typedef FunctionCall(Map<String, dynamic> params);
 
 /// reflects functions declared in a Dart program.
 class FunctionMirror {
-  final Function call;
+  final FunctionCall call;
   /*Type | List<Type, Type | List<Type> | List<Type, ...>>*/ final returnType;
   final Map<String, DeclarationMirror> parameters;
 
@@ -58,5 +73,9 @@ class DeclarationMirror {
   /*Type | List<Type, Type | List<Type> | List<Type, ...>>*/ final type;
   final List<Annotation> annotations;
 
-  const DeclarationMirror({this.type, this.annotations});
+  final isFinal;
+
+  final isOptional;
+
+  const DeclarationMirror({this.type, this.annotations, this.isFinal = false, this.isOptional = false});
 }